BITS0000: sailfish - log.New(...) -> logger.With(logger, ...)
The log interface has an issue that it is defined recursively and cannot be
implemented by any other module, which means we cant replace it. First step is
to stop using the .New() function defined in the interface.

We'll use "log.With()" as a replacement for .New(), they function the same.

Next step is that we remove the .New() from the interface.

All the above is implemented in this patch

After that, we should be able to switch out the backend implementation, if
needed. Logging implementations have a pretty big difference in number of
allocations and speed, so this is a pretty sensitive area where we could get
some good wins by subbing out.

log15 is what we are using now. Either zap or go-kit/log would be good
replacements, need more work to benchmark.

Change-Id: Ib6a4ccb49e4a8862382bf45aa925168959284384