ECNW-0: update redfish-ec.yaml to change fan slot 'name' to 'fanname' mapping
Both FanEvent and AttributeUpdated contain the name of the fan as "FanName".
However, for AttributeUpdated, this gets mapped to "Name" inside the cfg which
then differs from the FanEvent. By fixing the mapping to stay as "fanname", both
FanEvent and AttributeUpdated will be able to set the fan name. When merged up
to 1.3, this will keep the code the same since FanEvent is removed.

Change-Id: If14ed157684c07e4da81cceca48dbc18a589db49