JIT-164366: instantiate fwupdate tasks in a seperate GO routine
When instantiating any new URIs for a fwupdate task group, if the attributes
were recycled/used too quickly, they would be missed if the routine was stuck
instantiating a previous task. By putting the instantiate in its own routine,
any new changes can be processed much quicker and will be far less likely to be
missed.

Change-Id: I97d116d45d575be3631981baa9bf16facb2eb119